requirements.txt
Pillow
pypresence
pystray
Requests
